refactor:子供の情報に関する部分をクラスに分割
